+++ js/utils.js
@@ -1,3 +1,5 @@
+const ALERT_SHOW_TIME = 5000;
+
 //Нахождени случайного целого числа из диапозона включительно
 function randomInteger(min, max) {
   if (min < 0 || max <= min) {
@@ -29,4 +31,25 @@ let randomArray = function (arr) {
   return newarr;
 };
 
-export {randomInteger, randomFloat, randonElementArray, randomArray};
+const showAlert = (message) => {
+  const alertContainer = document.createElement('div');
+  alertContainer.style.zIndex = 100;
+  alertContainer.style.position = 'absolute';
+  alertContainer.style.left = 0;
+  alertContainer.style.top = 0;
+  alertContainer.style.right = 0;
+  alertContainer.style.padding = '10px 3px';
+  alertContainer.style.fontSize = '30px';
+  alertContainer.style.textAlign = 'center';
+  alertContainer.style.backgroundColor = 'red';
+
+  alertContainer.textContent = message;
+
+  document.body.append(alertContainer);
+
+  setTimeout(() => {
+    alertContainer.remove();
+  }, ALERT_SHOW_TIME);
+}
+
+export {randomInteger, randomFloat, randonElementArray, randomArray, showAlert};
